pyproject.toml
[project]
name = "openmeteo-exporter"
version = "1.0.0"
description = "Prometheus exporter for Open-Meteo weather data"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.9"
dependencies = [
    "prometheus-client>=0.19.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
